added postgres driver, updated racer repository and server socket handling
Took 57 minutes

+++ database source/pro2.sql
@@ -19,6 +19,6 @@ create table sep2.admin
 create table sep2.horse(
     id  serial primary key,
     name varchar,
-    speedMin float,
-    speedMax float
+    speedMin int,
+    speedMax int
 )
